+++ oneflow/user/kernels/upsample_linear_1d_kernel.cu
@@ -84,7 +84,7 @@ class UpsampleLinear1DGPUKernel final : public user_op::OpKernel {
                                                y_tensor->shape().At(2));
     const int64_t in_height = x_tensor->shape().At(2);
     const int64_t out_height = y_tensor->shape().At(2);
-    const float height_scale = out_height * 1.0 / in_height;
+    const double height_scale = out_height * 1.0 / in_height;
     if (in_height == out_height) {
       Memcpy<DeviceType::kCUDA>(
           ctx->stream(), y_tensor->mut_dptr<void>(), x_tensor->dptr<void>(),
@@ -121,7 +121,7 @@ class UpsampleLinearGrad1DGPUKernel final : public user_op::OpKernel {
     const int64_t elem_cnt = dy_tensor->shape().elem_cnt();
     const int64_t in_height = dx_tensor->shape().At(2);
     const int64_t out_height = dy_tensor->shape().At(2);
-    const float height_scale = out_height * 1.0 / in_height;
+    const double height_scale = out_height * 1.0 / in_height;
     if (in_height == out_height) {
       Memcpy<DeviceType::kCUDA>(
           ctx->stream(), dx_tensor->mut_dptr<void>(), dy_tensor->dptr<void>(),
